backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
pydub
openai
supabase
aiofiles
